Add script to remove marc:languageNote
